# docs/openapi-gateway.yaml
openapi: 3.1.0
info:
  title: snitchr Gateway API
  version: 0.1.0
  description: |
    API for the snitchr monitoring gateway. Runs on Cloudflare Workers, Node.js, or Bun.

    Two auth models:
    - **Session cookie** — HMAC-signed expiry timestamp in httpOnly cookie (24h TTL). Used by all dashboard/admin endpoints.
    - **Bearer token** — per-machine 48-char hex secret. Used by the agent ingest endpoint only.
  license:
    name: MIT
    url: https://github.com/thinking-tools/snitchr/blob/main/LICENSE
  contact:
    name: snitchr
    url: https://snitchr.sh
servers:
  - url: https://{host}
    description: snitchr gateway
    variables:
      host:
        default: localhost:8787
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: HMAC-signed expiry timestamp (24h TTL)
    bearerToken:
      type: http
      scheme: bearer
      description: Per-machine 48-char hex secret

  # ── Reusable schemas ──────────────────────────────────────────────────
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
    Ok:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
          enum: [true]
    EventLevel:
      type: integer
      enum: [0, 1, 2, 3, 4, 5]
      description: |
        0=HEARTBEAT, 1=INFO, 2=WARN, 3=ERROR, 4=CRITICAL, 5=SHUTDOWN
    EventType:
      type: string
      enum:
        - ssh_login
        - ssh_fail
        - sudo
        - su
        - auth_fail
        - new_process
        - new_port
        - mount_change
        - cpu_spike
        - mem_spike
        - disk_high
        - file_change
        - crontab_change
        - shutdown
    StorageMode:
      type: string
      enum: [r2, s3, filesystem, memory]
    PortInfo:
      type: object
      required: [port, addr, proc]
      additionalProperties: false
      properties:
        port:
          type: integer
        addr:
          type: string
        proc:
          type: string
    MountInfo:
      type: object
      required: [src, target, fs]
      additionalProperties: false
      properties:
        src:
          type: string
        target:
          type: string
        fs:
          type: string
    TtyInfo:
      type: object
      required: [user, tty, from, login]
      additionalProperties: false
      properties:
        user:
          type: string
        tty:
          type: string
        from:
          type: string
        login:
          type: number
    HeartbeatData:
      type: object
      required: [ts, up, cpu, mem, disk, load, procs, net, ports, users]
      additionalProperties: false
      properties:
        ts:
          type: number
          description: Timestamp (ms since epoch)
        up:
          type: number
          description: Uptime in seconds
        cpu:
          type: number
          description: CPU usage percentage
        mem:
          type: array
          items:
            type: number
          minItems: 2
          maxItems: 2
          description: '[used_mb, total_mb]'
        disk:
          type: array
          items:
            type: number
          description: Disk usage percentages per mount
        load:
          type: array
          items:
            type: number
          minItems: 3
          maxItems: 3
          description: '[1min, 5min, 15min] load averages'
        procs:
          type: integer
          description: Running process count
        net:
          type: array
          items:
            type: number
          minItems: 2
          maxItems: 2
          description: '[rx_bytes, tx_bytes] cumulative'
        ports:
          type: string
          description: Comma-separated listening ports
        users:
          type: string
          description: Comma-separated logged-in users
        procList:
          type: string
          description: Pipe-delimited unique process names (sent on change only)
        portList:
          type: array
          items:
            $ref: '#/components/schemas/PortInfo'
          description: Detailed listening ports (sent on change only)
        mountList:
          type: array
          items:
            $ref: '#/components/schemas/MountInfo'
          description: Mounted filesystems (sent on change only)
        ttyList:
          type: array
          items:
            $ref: '#/components/schemas/TtyInfo'
          description: Active TTY sessions (sent on change only)
    AlertData:
      type: object
      required: [type, msg]
      additionalProperties: false
      properties:
        type:
          $ref: '#/components/schemas/EventType'
        msg:
          type: string
          maxLength: 2000
        unit:
          type: string
    EventEntry:
      type: object
      required: [t, ts, d]
      properties:
        t:
          $ref: '#/components/schemas/EventLevel'
        ts:
          type: number
        d:
          oneOf:
            - $ref: '#/components/schemas/HeartbeatData'
            - $ref: '#/components/schemas/AlertData'
    DiskPrediction:
      type: object
      required: [trend]
      properties:
        trend:
          type: string
          enum: [filling, stable, shrinking]
        hoursToThreshold:
          type: ['number', 'null']
...
    MachineListItem:
      type: object
      required: [id, online, alertCount]
      description: Optional fields (label, lat, lng, city, country, lastSeen, d, diskPrediction) are omitted when not available, not null.
      properties:
        id:
          type: string
          format: uuid
        label:
          type: string
        registeredAt:
          type: number
        lat:
          type: number
        lng:
          type: number
        city:
          type: string
        country:
          type: string
        online:
          type: boolean
        lastSeen:
          type: number
        d:
          $ref: '#/components/schemas/HeartbeatData'
        alertCount:
          type: integer
        diskPrediction:
          $ref: '#/components/schemas/DiskPrediction'
...
  responses:
    Unauthorized:
      description: Invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

# ── Paths ─────────────────────────────────────────────────────────────
paths:
...
  /api/config:
    get:
      tags: [Setup & Auth]
      summary: Get dashboard config
      description: Returns machine IDs, alert threshold, and storage mode.
      operationId: getConfig
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Dashboard config
          content:
            application/json:
              schema:
                type: object
                required: [machineIds, alertThreshold, storageMode]
                properties:
                  machineIds:
                    type: array
                    items:
                      type: string
                      format: uuid
                  alertThreshold:
                    $ref: '#/components/schemas/EventLevel'
                  storageMode:
                    $ref: '#/components/schemas/StorageMode'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ── Agent Ingest ────────────────────────────────────────────────────
  /m/{id}/ingest:
    post:
      tags: [Agent Ingest]
      summary: Agent telemetry ingest
      description: |
        Primary agent endpoint. Accepts heartbeats (t=0) and alert events (t=1-5).

        **Side effects:**
        - t=0: updates machine status (metrics, lastSeen, disk velocity samples)
        - t=1-4: appends to event log and alerts array, triggers notifications if threshold met
        - t=5: marks machine offline immediately

        Events for disabled features are silently accepted (200/ok) without storage.
        Duplicate alerts (same type+msg within heartbeat window) are deduplicated.
      operationId: ingest
      security:
        - bearerToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [t, d]
              additionalProperties: false
              properties:
                t:
                  $ref: '#/components/schemas/EventLevel'
                d:
                  oneOf:
                    - $ref: '#/components/schemas/HeartbeatData'
                    - $ref: '#/components/schemas/AlertData'
              example:
                t: 0
                d:
                  ts: 1711929600000
                  up: 86400
                  cpu: 12.5
                  mem: [4096, 8192]
                  disk: [45]
                  load: [1.2, 0.8, 0.6]
                  procs: 142
                  net: [1073741824, 536870912]
                  ports: '22,80,443'
                  users: root
        description: Max 8 KB payload
      responses:
        '200':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          description: Payload too large (>8 KB)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Not configured or storage unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Machines ────────────────────────────────────────────────────────
  /api/machines:
    get:
      tags: [Machines]
      summary: List all machines
      description: Returns all registered machines with their latest status, alert counts, and disk predictions.
      operationId: listMachines
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Machine list
          content:
            application/json:
              schema:
                type: object
                required: [machines, heartbeatTimeout, alertThreshold]
                properties:
                  machines:
                    type: array
                    items:
                      $ref: '#/components/schemas/MachineListItem'
                  heartbeatTimeout:
                    type: integer
                    description: Seconds before a machine is considered offline (default 120)
                  alertThreshold:
                    $ref: '#/components/schemas/EventLevel'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/machines/{id}:
    get:
      tags: [Machines]
      summary: Get machine details
      description: Returns machine info, latest heartbeat data, pending alerts, and disk prediction.
      operationId: getMachine
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Machine details
          content:
            application/json:
              schema:
                type: object
                required: [id, online, alerts]
                description: Optional fields (label, city, country, lastSeen, d, diskPrediction) are omitted when not available.
                properties:
                  id:
                    type: string
                    format: uuid
                  label:
                    type: string
                  registeredAt:
                    type: number
                  city:
                    type: string
                  country:
                    type: string
                  online:
                    type: boolean
                  lastSeen:
                    type: number
                  d:
                    $ref: '#/components/schemas/HeartbeatData'
                  alerts:
                    type: array
                    items:
                      $ref: '#/components/schemas/EventEntry'
                  diskPrediction:
                    $ref: '#/components/schemas/DiskPrediction'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Machines]
      summary: Update machine
      description: Update machine label. Empty string clears the label.
      operationId: updateMachine
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                label:
                  type: string
                  maxLength: 100
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Machines]
      summary: Delete machine
      description: Deregister machine and purge all stored data (status, logs, alerts).
      operationId: deleteMachine
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/machines/{id}/events:
    get:
      tags: [Machines]
      summary: Get machine event log
      description: Paginated event history for a specific date. Returns newest first.
      operationId: getMachineEvents
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: date
          in: query
          description: Date filter (YYYY-MM-DD). Defaults to today.
          schema:
            type: string
            format: date
        - name: limit
          in: query
          description: Max events to return (1-200, default 50)
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: cursor
          in: query
          description: Pagination cursor from previous response
          schema:
            type: string
      responses:
        '200':
          description: Event list
          content:
            application/json:
              schema:
                type: object
                required: [events, truncated]
                properties:
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/EventEntry'
                  cursor:
                    type: string
                    description: Pass to next request for pagination
                  truncated:
                    type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/machines/{id}/alerts:
    get:
      tags: [Machines]
      summary: Get machine alerts
      description: Returns recent alerts (max 100, FIFO, newest first).
      operationId: getMachineAlerts
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Alert list
          content:
            application/json:
              schema:
                type: object
                required: [alerts]
                properties:
                  alerts:
                    type: array
                    items:
                      $ref: '#/components/schemas/EventEntry'
                    maxItems: 100
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Machines]
      summary: Clear machine alerts
      description: Dismiss all pending alerts for a machine.
      operationId: clearMachineAlerts
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Cleared
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Settings ────────────────────────────────────────────────────────
...